requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.8.4
joblib==1.4.2
python-dotenv==1.0.1
